pet-emporio/services/auth-service/requirements.txt
# pe-common shared library (local)
-e ../../packages/pe-common

# Web framework
fastapi>=0.111
uvicorn[standard]>=0.29

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

# Cache
redis[asyncio]>=5.0

# Auth
pyjwt[cryptography]>=2.8
passlib[bcrypt]>=1.7
pyotp>=2.9

# HTTP client
httpx>=0.27

# Config & logging
pydantic-settings>=2.0
structlog>=24